#220329 Improve performance for `icmaa-category` list component (#534)
We had a reactive component for each category-list-item. This caused a large latency on component destroyal when leaving the page and switching the cluster (1000-2000ms). I fixed it by removing the item-component and deliver it as static HTML.  
I also tried to implement a lazyloader like `vue-virtual-scroller` but this was a lot of overhead for this unique feature.